GSYE-880: Rename heatpump tank modules
