+++ src/gsy_e/models/strategy/energy_parameters/heatpump/virtual_heatpump_tank.py
@@ -9,12 +9,12 @@ from gsy_e.models.strategy.energy_parameters.heatpump.constants import (
     WATER_SPECIFIC_HEAT_CAPACITY,
     WATER_DENSITY,
 )
-from gsy_e.models.strategy.state.heatpump_tank_states.heatpump_all_tanks_state import AllTanksState
+from gsy_e.models.strategy.state.heatpump_tank_states.all_tanks_state import AllTanksState
 from gsy_e.models.strategy.energy_parameters.heatpump.tank_parameters import (
     TankParameters,
     HeatpumpTankTypes,
 )
-from gsy_e.models.strategy.state.heatpump_tank_states.heatpump_water_tank_state import (
+from gsy_e.models.strategy.state.heatpump_tank_states.water_tank_state import (
     WaterTankState,
 )
 from gsy_e.models.strategy.energy_parameters.heatpump.virtual_heatpump_solver import (
