+++ src/gsy_e/models/strategy/state/heatpump_tank_states/all_tanks_state.py
@@ -6,8 +6,8 @@ from gsy_framework.constants_limits import FLOATING_POINT_TOLERANCE, GlobalConfi
 from gsy_framework.utils import convert_kJ_to_kWh
 from pendulum import DateTime
 
-from gsy_e.models.strategy.state.heatpump_tank_states.heatpump_pcm_tank_state import PCMTankState
-from gsy_e.models.strategy.state.heatpump_tank_states.heatpump_water_tank_state import (
+from gsy_e.models.strategy.state.heatpump_tank_states.pcm_tank_state import PCMTankState
+from gsy_e.models.strategy.state.heatpump_tank_states.water_tank_state import (
     WaterTankState,
     TankStateBase,
 )
