+++ tests/strategies/energy_parameters/test_tank_energy_parameters.py
@@ -6,7 +6,7 @@ from pendulum import datetime
 from gsy_framework.constants_limits import GlobalConfig
 
 from gsy_e.models.strategy.energy_parameters.heatpump.tank_parameters import TankParameters
-from gsy_e.models.strategy.state.heatpump_tank_states.heatpump_all_tanks_state import AllTanksState
+from gsy_e.models.strategy.state.heatpump_tank_states.all_tanks_state import AllTanksState
 
 from gsy_e.models.strategy.energy_parameters.heatpump.virtual_heatpump_tank import (
     VirtualHeatpumpAllTanksState,
